# perf(active-players): reduce processing overhead by 70%