--- a/public/js/activePlayersBootstrap.js
+++ b/public/js/activePlayersBootstrap.js
@@ -105,10 +105,10 @@ var ActivePlayersViewModel = function () {
     self.isLoading(true);
     $.post('/request/user/list-currently-active.php')
       .done(function (data) {
-        self.players([]);
-        data.forEach((player) => {
-          self.players.push(self.ConvertToObservablePlayer(player));
-        });
+        self.players(data.reduce(function (result, player) {
+          return result.concat(self.ConvertToObservablePlayer(player));
+        }, []));
+
         self.lastUpdate(new Date());
         self.hasError(false);
         self.isLoading(false);
@@ -119,8 +119,21 @@ var ActivePlayersViewModel = function () {
       });
   };
 
-  this.RefreshActivePlayers();
-  setInterval(this.RefreshActivePlayers, 5000 * 60);
+  this.init = function () {
+    const FIVE_MINUTES = 5000 * 60;
+
+    this.RefreshActivePlayers();
+    setInterval(this.RefreshActivePlayers, FIVE_MINUTES);
+  };
+
+  // This check will fail in Safari.
+  if ('requestIdleCallback' in window) {
+    requestIdleCallback(() => {
+      this.init();
+    });
+  } else {
+    this.init();
+  }
 };
 
 ko.applyBindings(new ActivePlayersViewModel(), document.getElementById('active-players-component'));